Smart home key
This goes makes the home key go to the beginning of the current line
without the whitespace when it is pressed the first time and
when pressed a second time it goes to the beginning of the line.

Code review requested on the start of the line finding code. Is there a
more efficient way?

Actually addresses ninja-ide/ninja-ide#820