Check for _mini
In the editor.Editor.set_tab_usage definition, the code would check to see if self._mini evaluates to True.  This was a problem if self._mini had not been created yet, such as when opening a new file.  As I'm working on a Go project right now and don't feel like switching editors for it, I had to fix that in order to be able to open files in that project - because Go's style guide recommends tabs (and, in fact, gofmt converts spaces to tabs).